Added "Align Knots" level plus additional tweaks.
In the "Align Knots" level, the player packs the crate with the ideal number of rats when stopping the packing operation at the moment when the knot in the left rope is at the same height as the knot in the right rope.
Sped up game play and removed the warning message telling the player to await the next crate.