Target number for rats packed per crate is now the same for the Wall Mark level and Knots level.
The ideal stopping point in the Knots level is now the same as in the Wall Mark level.

The "Missed By" field in the Knots level used to indicate the difference between knot centers. Now that we are counting by how many rats the packer missed, the field reflects this change.

Fixed a word wrap error on the crates that came about because there are now more rats packed per crate.