Chainsaw (v2012-11-27) now has a user-editable 'Rank' game attribute, which is initially blank, but user can click on cells to edit any value (using latest DG).
Chainsaw now displays "Score:  ?", and creates game attributes "Score" which is initially blank.
When the user creates a formula for "Score" in DG, it appears in the Score column, and the next time that they play a game, the end of game report in Chainsaw shows the numeric or text score value instead of "?".