Made practice checkbox reset to its last state when 'cancel' is pressed. (fixes #210)
This also fixes a bug where the timer would count down when the game was
in practice, and the user unchecked practice and hit cancel. The level
was not changed, but the practice variable was, so it affected the timer
which used that variable.

Also added practice to save/restore state methods to make sure games
saved in practice mode are restored in practice mode.